pyproject.toml
[project]
name = "notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
]
